Expose Dev Console url in Server response
Fix issue where the payload for websocket messages was returning the extension development url incorrectly when connecting to the websocket on localhost. This was hardcoded to return the public url which errors out when the public url hasn't been set. We should be consistent and return the current requested url like we do for the JSON response.
Remove trailing slash on the apiRoot

Clean up tests so that configured values like the apiRoot and extension uuid is not hardcoded in the test. This makes it easier to distinguish between values we expect to be passed in externally vs. hardcoded ones.

Expose Dev Console url in ExtensionsResponse TS interface

Bump Api Service version to 4

Move API Service configs to a single place