Fix issue where partial extension updates was forcing boolean values to be false
There is a quirk in Go where a boolean that has not been initialized
is equivalent to false. When an extension update event comes in
with partial data as all boolean values that have not been initialized
are treated as false. This causes a bug where true boolean values are
overwritten by non-initialized boolean and set to false.

This is fixed by using a pointer to a boolean to distinguish between a
boolean that has been set to false vs. one that has not been set.