Fix stale localization issue
The existing merge logic is meant to work by merging partial data
with full data when an update event is triggered. This doesn't work
for localization data when a translation key is deleted as the merging
logic will always hold on to the previous values.

This is fixed by adding a custom transformer rule when merging localization
data so that it does a full replace instead of a merge with existing data.

This PR also does some clean ups by moving all localization logic into
its own file and making sure only the build process is responsible for
updating build related info instead of making some of that live in the
API.